refactor(analytics): decompose dry-run tooling
